VivitriBack/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
pinecone-client
sentence-transformers
python-multipart
pytesseract
pdf2image
pillow
numpy
requests